Updated Jekyll and Bootstrap.
Was forced into this, as the old `jekyll-assets` plugin no longer works
on newer OSes, and I couldn't even build the site.

The Bootstrap version change is resulting in a bunch of mostly-minor
styling differences, which I'm not _thrilled_ about but can live with.
I'll probably tweak things more after this, but this is definitely a
reasonable stopping point.